Clarify README.md
The changes should make it easier for a newcomer to integrate HIE with Vim. The previous version made me think that I as  Vim user had to clone the LanguageClient-neovim repo, when in reality vim-plug works just fine to install the plugin. Using vim-plug is arguably easier than cloning the repo and adding it to the runtime path.